Update path.exists to return relevant errors from fs.stat
The callback for path.exists now accepts two arguments, an error and
a boolean value which performs the same as before. The error is only
passed from fs.stat if it is something other than 'not found'.

This brings the callback in-line with the convention of returning an
error as the first argument, and allows the identification of errors
which stop the existence of the file from being determined (such as
permissions errors).

path.existsSync has also been updated, and throws whenever an error
would have been returned to the callback.